whistle-server/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
Deprecated==1.2.14
Django==5.0.6
djangorestframework==3.15.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
keycove==0.3.5
kombu==5.3.7
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
redis==5.0.7
requests==2.32.3
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
svix==1.24.0
twilio==9.2.2
types-Deprecated==1.2.9.20240311
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.2
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
